Refactor prometheus cleanup to be more robust
 - add timeout to locks, so that we fail more gracefully with sentry
 report

 - only enable multiprocess mode if > 1 worker, or env var explictly set

 - better handling of locking with gevent/eventlet workers

 - better logging in case of lock/cleanup failures
