+++ talisker/metrics.py
@@ -29,41 +29,18 @@ from __future__ import absolute_import
 
 from builtins import *  # noqa
 
-from collections import defaultdict, OrderedDict
 from contextlib import contextmanager
 import functools
 import logging
-import json
-import os
-import tempfile
 import time
 
-from talisker import prometheus_lock
 import talisker.statsd
-from talisker.util import pkg_version, TaliskerVersionException
 
 
 try:
     import prometheus_client
-    from prometheus_client.multiprocess import (
-        MultiProcessCollector,
-        mark_process_dead,
-    )
-    prom_version = pkg_version('prometheus_client')
-    if prom_version in ('0.4.0', '0.4.1'):
-        raise TaliskerVersionException(
-            'prometheus_client {} has a critical bug in multiprocess mode, '
-            'and is not supported in Talisker. '
-            'https://github.com/prometheus/client_python/issues/322'.format(
-                prom_version,
-            )
-        )
-
 except ImportError:
     prometheus_client = False
-
-    def mark_process_dead(pid):
-        pass
 
 try:
     import statsd
@@ -168,204 +145,3 @@ class Counter(Metric):
             client = talisker.statsd.get_client()
             name = self.get_statsd_name(labels)
             client.incr(name, amount)
-
-
-histogram_archive = 'histogram_archive.db'
-counter_archive = 'counter_archive.db'
-
-
-def histogram_sorter(sample):
-    # sort histogram samples in order of bucket size
-    name, labels, _ = sample[:3]
-    return name, float(labels.get('le', 0))
-
-
-def _filter_exists(paths):
-    exists = []
-    for path in paths:
-        if os.path.exists(path):
-            exists.append(path)
-    return exists
-
-
-def prometheus_cleanup_worker(pid):
-    """Aggregate dead worker's metrics into a single archive file."""
-    mark_process_dead(pid)
-    prom_dir = os.environ['prometheus_multiproc_dir']
-    worker_files = [
-        'histogram_{}.db'.format(pid),
-        'counter_{}.db'.format(pid),
-    ]
-    paths = _filter_exists(os.path.join(prom_dir, f) for f in worker_files)
-
-    # check at least one worker file exists
-    if not paths:
-        return
-
-    histogram_path = os.path.join(prom_dir, histogram_archive)
-    counter_path = os.path.join(prom_dir, counter_archive)
-    archive_paths = _filter_exists([histogram_path, counter_path])
-
-    collect_paths = paths + archive_paths
-    collector = MultiProcessCollector(None)
-
-    try:
-        metrics = collector.merge(collect_paths, accumulate=False)
-    except AttributeError:
-        metrics = legacy_collect(collect_paths)
-
-    tmp_histogram = tempfile.NamedTemporaryFile(delete=False)
-    tmp_counter = tempfile.NamedTemporaryFile(delete=False)
-    write_metrics(metrics, tmp_histogram.name, tmp_counter.name)
-
-    # ensure reader does get partial state
-    with prometheus_lock:
-        os.rename(tmp_histogram.name, histogram_path)
-        os.rename(tmp_counter.name, counter_path)
-
-        for path in paths:
-            os.unlink(path)
-
-
-def legacy_collect(files):
-    """This almost verbatim from MultiProcessCollector.collect(), pre 0.4.0
-
-    The original collects all results in a format designed to be scraped. We
-    instead need to collect limited results, in a format that can be written
-    back to disk. To facilitate this, this version of collect() preserves label
-    ordering, and does not aggregate the histograms.
-
-    Specifically, it differs from the original:
-
-    1. it takes its files as an argument, rather than hardcoding '*.db'
-    2. it does not accumulate histograms
-    3. it preserves label order, to facilitate being inserted back into an mmap
-       file.
-
-    It needs to be kept up to date with changes to prometheus_client as much as
-    possible, or until changes are landed upstream to allow reuse of collect().
-    """
-    from prometheus_client import core
-    metrics = {}
-    for f in files:
-        if not os.path.exists(f):
-            continue
-        # verbatim from here...
-        parts = os.path.basename(f).split('_')
-        typ = parts[0]
-        d = core._MmapedDict(f, read_mode=True)
-        for key, value in d.read_all_values():
-            # Note: key format changed in 0.4+
-            metric_name, name, labelnames, labelvalues = json.loads(key)
-
-            metric = metrics.get(metric_name)
-            if metric is None:
-                metric = core.Metric(metric_name, 'Multiprocess metric', typ)
-                metrics[metric_name] = metric
-
-            if typ == 'gauge':
-                pid = parts[2][:-3]
-                metric._multiprocess_mode = parts[1]
-                metric.add_sample(
-                    name,
-                    tuple(zip(labelnames, labelvalues)) + (('pid', pid), ),
-                    value,
-                )
-            else:
-                # The duplicates and labels are fixed in the next for.
-                metric.add_sample(
-                    name,
-                    tuple(zip(labelnames, labelvalues)),
-                    value,
-                )
-        d.close()
-
-    for metric in metrics.values():
-        samples = defaultdict(float)
-        buckets = {}
-        for name, labels, value in metric.samples:
-            if metric.type == 'gauge':
-                without_pid = tuple(l for l in labels if l[0] != 'pid')
-                if metric._multiprocess_mode == 'min':
-                    current = samples.setdefault((name, without_pid), value)
-                    if value < current:
-                        samples[(name, without_pid)] = value
-                elif metric._multiprocess_mode == 'max':
-                    current = samples.setdefault((name, without_pid), value)
-                    if value > current:
-                        samples[(name, without_pid)] = value
-                elif metric._multiprocess_mode == 'livesum':
-                    samples[(name, without_pid)] += value
-                else:  # all/liveall
-                    samples[(name, labels)] = value
-
-            elif metric.type == 'histogram':
-                bucket = tuple(float(l[1]) for l in labels if l[0] == 'le')
-                if bucket:
-                    # _bucket
-                    without_le = tuple(l for l in labels if l[0] != 'le')
-                    buckets.setdefault(without_le, {})
-                    buckets[without_le].setdefault(bucket[0], 0.0)
-                    buckets[without_le][bucket[0]] += value
-                else:
-                    # _sum/_count
-                    samples[(name, labels)] += value
-
-            else:
-                # Counter and Summary.
-                samples[(name, labels)] += value
-
-        # end of verbatim copy
-        # modified to remove accumulation
-        if metric.type == 'histogram':
-            for labels, values in buckets.items():
-                for bucket, value in sorted(values.items()):
-                    key = (
-                        metric.name + '_bucket',
-                        labels + (('le', core._floatToGoString(bucket)),),
-                    )
-                    samples[key] = value
-
-        # Convert to correct sample format.
-        metric.samples = [
-            # OrderedDict used instead of dict
-            (name, OrderedDict(labels), value)
-            for (name, labels), value in samples.items()
-        ]
-    return metrics.values()
-
-
-def write_metrics(metrics, histogram_file, counter_file):
-    from prometheus_client.core import _MmapedDict
-    try:
-        from prometheus_client.core import _mmap_key
-    except ImportError:
-        # pre 0.4 key format
-        def _mmap_key(metric_name, name, labelnames, labelvalues):
-            return json.dumps(
-                (metric_name, name, tuple(labels), tuple(labels.values()))
-            )
-
-    histograms = _MmapedDict(histogram_file)
-    counters = _MmapedDict(counter_file)
-
-    try:
-        for metric in metrics:
-            if metric.type == 'histogram':
-                sink = histograms
-            elif metric.type == 'counter':
-                sink = counters
-
-            for sample in metric.samples:
-                # prometheus_client 0.4+ adds extra fields
-                name, labels, value = sample[:3]
-                key = _mmap_key(
-                    metric.name,
-                    name,
-                    tuple(labels),
-                    tuple(labels.values()),
-                )
-                sink.write_value(key, value)
-    finally:
-        histograms.close()
-        counters.close()
