+++ tests/conftest.py
@@ -33,13 +33,14 @@ import os
 # make sure prometheus is setup in multiprocess mode. We don't actually use
 # this dir in tests, as each test gets it's own directory, but this ensures
 # prometheus_client is imported in multiprocess mode
-from talisker import setup_prometheus_multiproc
-setup_prometheus_multiproc()
+from talisker.prometheus import setup_prometheus_multiproc
+setup_prometheus_multiproc(async_mode=False)
 
 # do this as early as possible, to set up logging in pytest
 import talisker.logs
 talisker.logs.configure_test_logging()
 talisker.logs.supress_noisy_logs()
+talisker._flush_early_logs()
 
 from wsgiref.util import setup_testing_defaults
